bpl-web-backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
supabase==2.5.0
python-dotenv==1.0.1
httpx==0.27.0

# Testing libraries
pytest==7.4.3
pytest-cov==6.2.1
anyio==3.7.1
